bundle update actionview activesupport activemodel
My end goal is to update actionview because it has a XSS vulnerability in v7.0.2.3. But I can't update only actionview because it's locked into a specific version of activesupport. Also, I can't update only activesupport because activemodel is locked into a specific version of activesupport. So that's why I did a batch update of the three gems.